requirements.txt
opencv-python
pytesseract
openai
pillow